Updated VS Solution and CMakeLists.txt to reflect latest changes Added annotation to timer bug on windows Fixed a wrong visibility in the static build of libkqueue
git-svn-id: svn://svn.code.sf.net/p/libkqueue/code/trunk@489 fb4e3144-bc1c-4b72-a658-5bcd248dd7f7

+++ src/linux/read.c
@@ -49,7 +49,7 @@ get_eof_offset(int fd)
         sb.st_size = 1;
     }
 
-    dbg_printf("curpos=%zu size=%zu\n", curpos, sb.st_size);
+    dbg_printf("curpos=%zu size=%zu\n", (size_t)curpos, (size_t)sb.st_size);
     return (sb.st_size - curpos); //FIXME: can overflow
 }
